[ez] move addoption to conftest in root test directory and fix broken imports
