+++ llama_stack/providers/tests/eval/conftest.py
@@ -76,22 +76,6 @@ def pytest_configure(config):
         )
 
 
-def pytest_addoption(parser):
-    parser.addoption(
-        "--inference-model",
-        action="store",
-        default="meta-llama/Llama-3.2-3B-Instruct",
-        help="Specify the inference model to use for testing",
-    )
-
-    parser.addoption(
-        "--judge-model",
-        action="store",
-        default="meta-llama/Llama-3.1-8B-Instruct",
-        help="Specify the judge model to use for testing",
-    )
-
-
 def pytest_generate_tests(metafunc):
     if "eval_stack" in metafunc.fixturenames:
         available_fixtures = {
